+++ test/runtime/samples/dynamic-element-attribute/main.svelte
@@ -1,0 +1,5 @@
+<script>
+  export let tag = 'div';
+</script>
+
+<svelte:element this={tag} style="color: red;">Foo</svelte:element>