+++ test/runtime/samples/dynamic-element-change-tag/main.svelte
@@ -1,0 +1,5 @@
+<script>
+  export let tag = 'div';
+</script>
+
+<svelte:element this={tag}>Foo</svelte:element>