+++ test/validator/samples/dynamic-element-this/input.svelte
@@ -1,0 +1,3 @@
+<div>
+	<svelte:element {this}>foo</svelte:element>
+</div>